+++ CHANGELOG.md
@@ -35,6 +35,7 @@ A more detailed list of changes is available in the corresponding milestones for
   - **[com.google.fonts/check/039]:** com.google.fonts/check/fontforge
   - **[com.google.fonts/check/040]:** com.google.fonts/check/win_ascent_and_descent
   - **[com.google.fonts/check/041]:** com.google.fonts/check/linegaps
+  - **[com.google.fonts/check/042]:** com.google.fonts/check/os2_metrics_match_hhea
   - **[com.google.fonts/check/045]:** com.google.fonts/check/dsig
   - **[com.google.fonts/check/082]:** com.google.fonts/check/metadata/listed_on_gfonts
   - **[com.google.fonts/check/083]:** com.google.fonts/check/metadata_unique_full_name_values